# Optimize Enum.map/2 for lists