@josevalim 2015-10-01
Just to be double sure, please run every benchmark in a separated process.
:)

-- 


*José Valim*
www.plataformatec.com.br
Skype: jv.ptec
Founder and Director of R&D