@michalmuskala 2015-10-01
I must admin, and went a lazy path this time - [the new testing code](https://gist.github.com/michalmuskala/7feffb201219de0c6276)

Results: 
```
cons 100            500000   5.62 µs/op
recurse 100         500000   6.47 µs/op
for 100             200000   8.75 µs/op
cons 1000            50000   57.47 µs/op
recurse 1000         50000   59.15 µs/op
for 1000             50000   70.73 µs/op
cons 10000            5000   535.17 µs/op
recurse 10000         5000   573.46 µs/op
for 10000             5000   678.71 µs/op
cons 100000            200   7505.32 µs/op
recurse 100000         200   7814.68 µs/op
for 100000             200   9194.78 µs/op
```