@josevalim 2015-10-01
Yes. comprehensions have been rewritten to always use Enum.reduce, that's why it became slower. If you find any other use of comprehensions in `Enum`, we should change it. Can you please send PRs for any other cases you see? Thank you! :)