@splattael 2015-10-01
> Just to be double sure, please run every benchmark in a separated process.
@josevalim Interesting, could you explain why? Should we always run benchmarks in separated processes?