@michalmuskala 2015-10-01
I think the concern was to eliminate the effects of garbage collection. Each process has a new heap, so garbage collection will happen at the same time. When you run all in the same process it's not as predictable, as you can accumulate garbage from previous runs.