@splattael 2015-10-01
@michalmuskala That makes sense. Thank you for the explanation :)
`benchp` is pretty handy! Maybe it should be added to `Benchfella` :D