@ericmj 2015-10-03
I'm usually a proponent of body-recursive functions but in the stdlib we should strive for highest performance. @michalmuskala did you also measure the tail-recursive version that reverses the list at the end. Tail-recursive should be ~30% faster on x86 according to erlang's efficiency guide.