@michalmuskala 2015-10-03
@ericmj I believe tail-recursive is the implementation I called `recurse`. It's performance is similar to the body-recursive solution I called `cons` that I ended up using, because it came out as slightly faster.
I may be wrong on that. Maybe someone could run those tests on other machine that mine.

I've read that part of efficiency guide too - if my benchmarking was correct it seems both solutions are equally performant currently.