@lexmag 2015-10-03
I think using `:lists.map/2` makes it even better: similar performance and obviously less code.