@josevalim 2015-10-04
In my previous benchmarks, tail-recursive is faster for long collections, body-recursive for shorter ones. `:lists.map` is also great because we inline its code. They are all great solutions.